+++ src/main/java/frc/robot/util/SwerveModuleAcceleration.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.util;
 
 import java.util.Objects;
